+++ packages/costa/test/runtime_test.ts
@@ -3,7 +3,7 @@ import { CostaRuntime } from '../src/runtime';
 import { PluginPackage } from '../src';
 import { stat } from 'fs-extra';
 import { spawnSync } from 'child_process';
-import { createReadStream, createWriteStream, remove } from 'fs-extra';
+import { createReadStream } from 'fs-extra';
 
 const INSTALL_SPECS_TIMEOUT = 180 * 1000;
 
@@ -15,18 +15,6 @@ describe('create a costa runtime', () => {
     npmRegistryPrefix: 'https://registry.npmjs.com/'
   });
   let collectCsv: PluginPackage;
-
-  it('test upload file saving', async () => {
-    let step = 0;
-    const writeFile = path.join(__dirname, 'runtime.ts_test');
-    const readStream = createReadStream(path.join(__dirname, 'runtime.ts'));
-    const writeStream = createWriteStream(writeFile);
-    readStream.on('end', () => expect(++step).toBe(0));
-    writeStream.on('finish', () => expect(++step).toBe(1));
-    await costa.saveFile(readStream, writeStream);
-    expect(step).toBe(2);
-    await remove(writeFile);
-  }, INSTALL_SPECS_TIMEOUT);
 
   it('should fetch a plugin and install from local', async () => {
     collectCsv = await costa.fetch(path.join(process.cwd(), '../plugins/data-collect/csv-data-collect'));
