+++ packages/costa/test/utils_test.ts
@@ -1,0 +1,15 @@
+import path from 'path';
+import { createReadStream, createWriteStream, remove } from 'fs-extra';
+import { pipeWithFinish } from '../src/utils';
+
+it('test pipe with finish event', async () => {
+  let step = 0;
+  const writeFile = path.join(__dirname, 'utils_test.ts_test');
+  const readStream = createReadStream(path.join(__dirname, 'utils_test.ts'));
+  const writeStream = createWriteStream(writeFile);
+  readStream.on('end', () => expect(++step).toBe(0));
+  writeStream.on('finish', () => expect(++step).toBe(1));
+  await pipeWithFinish(readStream, writeStream);
+  expect(step).toBe(2);
+  await remove(writeFile);
+});
